// front-end/photo-dom/src/services/photo-service/models/photo-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PhotoDto {
  '@class': string;
  contentType?: string;
  filename?: string;
  id?: number;
  imageBase64?: string;
  imageBytes?: Array<string>;
  likeCount?: number;
  size?: number;
  userId?: number;
}